% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_geocarb.R
\name{check_dependencies}
\alias{check_dependencies}
\title{check_dependencies}
\usage{
check_dependencies()
}
\value{
Logical indicating success.
}
\description{
Check that a reasonable version of python is installed.
}
\keyword{internal}
