% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocarb-package.R
\docType{package}
\name{geocarb-package}
\alias{geocarb}
\alias{geocarb-package}
\title{geocarb: Simulates the geochemical carbon cycle using the GEOCARB model}
\description{
This package uses a modifier version of the GEOCARB model by 
  Robert Berner, as modified by David Archer and then translated into 
  python and further modified by Jonathan Gilligan.
}
\author{
\strong{Maintainer}: Jonathan M. Gilligan \email{jonathan.gilligan@vanderbilt.edu} (\href{https://orcid.org/0000-0003-1375-6686}{ORCID})

}
\keyword{internal}
