% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_geocarb.R
\name{load_geocarb}
\alias{load_geocarb}
\title{load_geocarb}
\usage{
load_geocarb(python_script = get_py_geocarb_path())
}
\arguments{
\item{python_script}{Which script to load.}
}
\value{
Logical indicating success.
}
\description{
Load the geocarb module.
}
