% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geocarb.R
\name{read_geocarb}
\alias{read_geocarb}
\title{read_geocarb}
\usage{
read_geocarb(filename)
}
\arguments{
\item{filename}{The file to read.}
}
\value{
A tibble containing the results of the GEOCARB simulation.
See \link{run_geocarb} for details.
}
\description{
Read a GEOCARB output file.
}
\seealso{
run_geocarb
}
